+++ android/src/main/kotlin/com/imagepicker/listeners/OnFolderClickListener.kt
@@ -1,7 +1,0 @@
-package com.imagepicker.listeners
-
-import com.imagepicker.model.Folder
-
-interface OnFolderClickListener {
-    fun onFolderClick(bucket: Folder)
-}