+++ android/src/main/kotlin/com/imagepicker/listeners/OnImageSelectedListener.kt
@@ -1,7 +1,0 @@
-package com.imagepicker.listeners
-
-import com.imagepicker.model.Image
-
-interface OnImageSelectedListener {
-    fun onSelectionUpdate(selectedImage: List<Image?>?)
-}