+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+- Upgrade gallery for Android
+
 ## 0.1.1
 
 - Upgrade using lasted flutter for iOS